增加readme，解决上传信息卡死问题

+++ pc-sdk/csharp/NexLinkNative.cs
@@ -27,7 +27,7 @@ namespace NexLink
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)]
         public static extern int nexlink_cmd(
             IntPtr handle,
-            byte cmd,
+            ushort cmd,
             byte[] payload,
             ushort length,
             out NexLinkPacket resp,
